fix signup login dialogs (backdrop works again), add reset functionality to forms
